requirements.txt
asgiref
django
Pillow
sqlparse
tzdata
django-ckeditor-5